Rename function parameter for clarity
The change in d753d4b made `bubble_down()` generic to be able to
bubble-down a value at an index not just at the root, but anywhere in
the heap. Therefore, the parameter name `newRoot` is misleading, as it
assumes that the index of the value being bubbled-down is the root
index, which is now no longer the case.