Replace dynamic array underneath BST with slice
Not many of properties of dynamic arrays will be used for the binary
search tree, so a slice was chosen instead.

For example, when the BST needs to be grown by adding an extra layer,
it'll always be by a specific amount rather than by one element at a
time, which is something the dynamic array would provide.

Furthermore, the BST should be able to have null nodes, so a slice of
`?u8` was chosen.