Enable inorder traversal to work with BST's with `null` nodes
Temporarily allow insertion of `null` values into the BST to make it
easier to setup an array-represention of the BST with null nodes (to
then be able to test the inorder traversal algorithm is correct).

The type of values to insert will change back to `u8`, and the `count`
field (added for conveniently being able to insert values into the BST),
will go away in the near future.