Replace slice underneath BST with linked nodes
Looking ahead, removal of elements in a BST with more than a root node
would be tricky with an array representation of the elements in the BST.
Mainly, needing to shift a left/right subtree of a node within the
underlying slice if the node is removed seemed fiddly.

So, an underlying structure of nodes connected to their left/right
children via pointers seemed better equipped to handle node removal (as
far as being simpler to implement, compared to if an array
representation was used).