Replace array of nodes with array of integers
The integer values in the array correspond to the "parent" of each
element that is tracked in the union-find. The representative of an
element will be found by following a "trail of indices" stored in an
array, rather than a "trail of parent pointers" of nodes that are stored
in an array.